requirements.txt
# requirements.txt
python-whois
tqdm
requests
colorama
dnspython
re
time
json